requirements.txt
absl-py==2.1.0
accelerate==0.23.0
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.2.0
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
autopep8==2.1.0
backoff==2.2.1
bar==0.2.1
bert-score==0.3.13
bitsandbytes==0.42.0
boilerpy3==1.0.6
Bottleneck==1.3.7
brotlipy==0.7.0
cachetools==5.3.3
cattrs==23.1.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cmake==3.28.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.1
cryptography==41.0.7
cycler==0.12.1
datasets==2.16.1
dill==0.3.7
docopt==0.6.2
dpcpp-cpp-rt==2024.0.2
elastic-transport==7.16.0
elasticsearch==7.17.9
evaluate==0.4.1
Events==0.5
exceptiongroup==1.1.3
executing==2.0.1
falcon==3.1.1
farm-haystack==1.23.0
filelock==3.13.1
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
future==0.18.3
google-auth==2.29.0
google-auth-oauthlib==1.0.0
grpcio==1.63.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.2
humanfriendly==10.0
icecream==2.1.3
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
inflect==7.0.0
intel-cmplr-lib-rt==2024.0.2
intel-cmplr-lic-rt==2024.0.2
intel-opencl-rt==2024.0.2
intel-openmp==2024.0.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
lazy-imports==0.3.1
lit==17.0.6
lxml==5.1.0
Markdown==3.6
MarkupSafe==2.1.3
matplotlib==3.7.4
mkl==2024.0.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
monotonic==1.6
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.1
nlp==0.4.0
nltk==3.8.1
num2words==0.5.12
numexpr==2.8.6
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
optimum==1.16.1
packaging==23.2
pandas==2.0.3
Pillow==10.0.1
pkgutil-resolve-name==1.3.10
platformdirs==3.10.0
portalocker==2.8.2
posthog==3.0.2
progress==1.6
progressbar==2.5
prompthub-py==4.0.0
protobuf==4.24.3
psutil==5.9.5
pyarrow==11.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==1.10.12
pygments==2.17.2
pyOpenSSL==23.3.0
pyparsing==3.1.1
pypdf2==3.0.1
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
quantulum3==0.9.0
rank-bm25==0.2.2
referencing==0.30.2
regex==2023.12.25
requests==2.31.0
requests-cache==0.9.8
requests-oauthlib==2.0.0
responses==0.18.0
rouge-score==0.1.2
rpds-py==0.10.3
rsa==4.9
sacrebleu==2.4.0
sacremoses==0.0.53
safetensors==0.4.1
scikit-learn==1.3.1
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
sseclient-py==1.8.0
sympy==1.12
tabulate==0.9.0
tbb==2021.11.0
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.2.0
tiktoken==0.5.1
tokenizers==0.14.1
tomli==2.0.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.1
transformers==4.36.0
triton==2.0.0
typing-extensions==4.9.0
tzdata==2023.4
url-normalize==1.4.3
urllib3==2.1.0
werkzeug==3.0.3
win-inet-pton==1.1.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
